move swizzling to internal module
