+++ Sources/MessagingPush/PushClickHandler.swift
@@ -88,52 +88,3 @@ extension PushClickHandlerImpl {
         cio_swizzle_didReceive(center, didReceive: response, withCompletionHandler: completionHandler)
     }
 }
-
-// TODO: cleanup duplicate swizzle functions all over codebase.
-
-// https://nshipster.com/swift-objc-runtime/
-// https://github.com/expo/expo/blob/ad56550cff90602645f215d5eebd53e59fe2df76/packages/expo-dev-launcher/ios/EXDevLauncherUtils.swift#L9-L24
-func swizzle(forClass: AnyClass, original: Selector, new: Selector) {
-    guard let originalMethod = class_getInstanceMethod(forClass, original) else { return }
-    guard let swizzledMethod = class_getInstanceMethod(forClass, new) else { return }
-
-    let didAddMethod = class_addMethod(forClass, original, method_getImplementation(swizzledMethod), method_getTypeEncoding(swizzledMethod))
-
-    if didAddMethod {
-        class_replaceMethod(forClass, new, method_getImplementation(originalMethod), method_getTypeEncoding(originalMethod))
-    } else {
-        method_exchangeImplementations(originalMethod, swizzledMethod)
-    }
-}
-
-// https://github.com/OneSignal/OneSignal-iOS-SDK/blob/5ac5927fca8361a3af9cee1262ac11fd49e218ee/iOS_SDK/OneSignalSDK/OneSignalCore/Source/Swizzling/OneSignalSelectorHelpers.m#L33
-func swizzle(targetClass: AnyClass, targetSelector: Selector, myClass: AnyClass, mySelector: Selector) {
-    // TODO: make this runtime safe.
-    guard let newMethod = class_getInstanceMethod(myClass, mySelector) else {
-        fatalError()
-//        return
-    }
-    let newImplementation = method_getImplementation(newMethod)
-
-    let methodTypeEncoding = method_getTypeEncoding(newMethod)
-
-    let existingMethod = class_getInstanceMethod(targetClass, targetSelector)
-    if existingMethod != nil {
-        guard let originalMethod = class_getInstanceMethod(targetClass, targetSelector) else {
-            fatalError()
-//            return
-        }
-        let originalImplementation = method_getImplementation(originalMethod)
-
-        guard newImplementation != originalImplementation else {
-            fatalError()
-//            return
-        }
-
-        class_addMethod(targetClass, mySelector, newImplementation, methodTypeEncoding)
-        let newMethod = class_getInstanceMethod(targetClass, mySelector)
-        method_exchangeImplementations(originalMethod, newMethod!)
-    } else {
-        class_addMethod(targetClass, targetSelector, newImplementation, methodTypeEncoding)
-    }
-}
