+++ Sources/MessagingPushAPN/MessagingPushAPN+PushConfigs.swift
@@ -1,3 +1,4 @@
+import CioInternalCommon
 import Foundation
 import UIKit
 
@@ -11,38 +12,34 @@ extension MessagingPushAPN {
 
     private func swizzleDidRegisterForRemoteNotifications() {
         let appDelegate = UIApplication.shared.delegate
-        let appDelegateClass: AnyClass? = object_getClass(appDelegate)
+        guard let appDelegateClass = object_getClass(appDelegate) else {
+            return
+        }
 
         // Swizzle `didRegisterForRemoteNotificationsWithDeviceToken`
         let originalSelector = #selector(UIApplicationDelegate.application(_:didRegisterForRemoteNotificationsWithDeviceToken:))
         let swizzledSelector = #selector(application(_:didRegisterForRemoteNotificationsWithDeviceToken:))
-        swizzle(forOriginalClass: appDelegateClass, forSwizzledClass: MessagingPushAPN.self, original: originalSelector, new: swizzledSelector)
+
+        swizzle(targetClass: appDelegateClass, targetSelector: originalSelector, myClass: MessagingPushAPN.self, mySelector: swizzledSelector)
 
         // Swizzle `didFailToRegisterForRemoteNotificationsWithError`
         let originalSelectorForDidFail = #selector(UIApplicationDelegate.application(_:didFailToRegisterForRemoteNotificationsWithError:))
         let swizzledSelectorForDidFail = #selector(application(_:didFailToRegisterForRemoteNotificationsWithError:))
-        swizzle(forOriginalClass: appDelegateClass, forSwizzledClass: MessagingPushAPN.self, original: originalSelectorForDidFail, new: swizzledSelectorForDidFail)
-    }
 
-    private func swizzle(forOriginalClass: AnyClass?, forSwizzledClass: AnyClass?, original: Selector, new: Selector) {
-        guard let swizzledMethod = class_getInstanceMethod(forSwizzledClass, new) else { return }
-        guard let originalMethod = class_getInstanceMethod(forOriginalClass, original) else {
-            // Add method if it doesn't exist
-            class_addMethod(forOriginalClass, new, method_getImplementation(swizzledMethod), method_getTypeEncoding(swizzledMethod))
-            return
-        }
-        method_exchangeImplementations(originalMethod, swizzledMethod)
+        swizzle(targetClass: appDelegateClass, targetSelector: originalSelectorForDidFail, myClass: MessagingPushAPN.self, mySelector: swizzledSelectorForDidFail)
     }
 
     // Swizzled method for APN device token.
-    @objc
-    func application(_ application: UIApplication, didRegisterForRemoteNotificationsWithDeviceToken deviceToken: Data) {
+    @objc dynamic func application(_ application: UIApplication, didRegisterForRemoteNotificationsWithDeviceToken deviceToken: Data) {
         Self.shared.registerDeviceToken(apnDeviceToken: deviceToken)
+
+        self.application(application, didRegisterForRemoteNotificationsWithDeviceToken: deviceToken) // continue swizzle
     }
 
     // Swizzled method for `didFailToRegisterForRemoteNotificationsWithError'
-    @objc
-    func application(_ application: UIApplication, didFailToRegisterForRemoteNotificationsWithError error: Error) {
+    @objc dynamic func application(_ application: UIApplication, didFailToRegisterForRemoteNotificationsWithError error: Error) {
         messagingPush.deleteDeviceToken()
+
+        self.application(application, didFailToRegisterForRemoteNotificationsWithError: error) // continue swizzle
     }
 }
